requirements.txt
asgiref==3.8.1
boto3==1.38.21
botocore==1.38.21
chardet==5.2.0
dj-database-url==2.3.0
Django==5.1.7
django-storages==1.14.6
djangorestframework==3.16.0
et_xmlfile==2.0.0
git-filter-repo==2.47.0
jmespath==1.0.1
Levenshtein==0.27.1
numpy==2.2.6
openpyxl==3.1.5
pandas==2.3.0
pillow==11.2.1
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-Levenshtein==0.27.1
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
reportlab==4.0.4
s3transfer==0.12.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
whitenoise==6.9.0
